fix datatype

+++ migrations/20240501015227_create_cats.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS cats(
   name VARCHAR (255) NOT NULL,
   race VARCHAR (255) NOT NULL,
   sex VARCHAR (10) NOT NULL,
-  age_in_month SMALLSERIAL,
+  age_in_month INT,
   description VARCHAR (255) NOT NULL,
   image_urls VARCHAR[] NOT NULL,
   user_id INT REFERENCES users(id) NOT NULL,
